bench script: handle lack of html/ directory gracefully

+++ dev/bench/bench.sh
@@ -542,10 +542,12 @@ timings=$working_dir/timings
 mkdir -p $timings
 
 # Print line by line slow downs and speed ups
+if [ -d "$working_dir/html" ]; then # might not exist if all jobs failed
 cd "$working_dir/html"
 $render_line_results
 # Move line timing files to timings folder (they will become artifacts)
 mv fast_table slow_table timings_table $timings
+fi
 
 echo "INFO: workspace = ${CI_JOB_URL}/artifacts/browse/${bench_dirname}"
